feat: init usage file
