+++ lib/api.ts
@@ -1,10 +1,6 @@
 import { UserProfile } from "@/app/types/user"
 import { APP_DOMAIN, DAILY_SPECIAL_AGENT_LIMIT } from "@/lib/config"
 import { SupabaseClient } from "@supabase/supabase-js"
-import {
-  AUTH_DAILY_MESSAGE_LIMIT,
-  NON_AUTH_DAILY_MESSAGE_LIMIT,
-} from "./config"
 import { fetchClient } from "./fetch"
 import { API_ROUTE_CREATE_GUEST, API_ROUTE_UPDATE_CHAT_MODEL } from "./routes"
 
@@ -39,148 +35,6 @@ export class UsageLimitError extends Error {
     super(message)
     this.code = "DAILY_LIMIT_REACHED"
   }
-}
-
-/**
- * Checks the user's daily usage to see if they've reached their limit.
- * Uses the `anonymous` flag from the user record to decide which daily limit applies.
- *
- * @param supabase - Your Supabase client.
- * @param userId - The ID of the user.
- * @throws UsageLimitError if the daily limit is reached, or a generic Error if checking fails.
- * @returns User data including message counts and reset date
- */
-export async function checkUsage(supabase: SupabaseClient, userId: string) {
-  const { data: userData, error: userDataError } = await supabase
-    .from("users")
-    .select(
-      "message_count, daily_message_count, daily_reset, anonymous, premium"
-    )
-    .eq("id", userId)
-    .maybeSingle()
-
-  if (userDataError) {
-    throw new Error("Error fetchClienting user data: " + userDataError.message)
-  }
-  if (!userData) {
-    throw new Error("User record not found for id: " + userId)
-  }
-
-  // Decide which daily limit to use.
-  const isAnonymous = userData.anonymous
-  // (Assuming these are imported from your config)
-  const dailyLimit = isAnonymous
-    ? NON_AUTH_DAILY_MESSAGE_LIMIT
-    : AUTH_DAILY_MESSAGE_LIMIT
-
-  // Reset the daily counter if the day has changed (using UTC).
-  const now = new Date()
-  let dailyCount = userData.daily_message_count || 0
-  const lastReset = userData.daily_reset ? new Date(userData.daily_reset) : null
-
-  if (
-    !lastReset ||
-    now.getUTCFullYear() !== lastReset.getUTCFullYear() ||
-    now.getUTCMonth() !== lastReset.getUTCMonth() ||
-    now.getUTCDate() !== lastReset.getUTCDate()
-  ) {
-    dailyCount = 0
-    const { error: resetError } = await supabase
-      .from("users")
-      .update({ daily_message_count: 0, daily_reset: now.toISOString() })
-      .eq("id", userId)
-    if (resetError) {
-      throw new Error("Failed to reset daily count: " + resetError.message)
-    }
-  }
-
-  // Check if the daily limit is reached.
-  if (dailyCount >= dailyLimit) {
-    throw new UsageLimitError("Daily message limit reached.")
-  }
-
-  return {
-    userData,
-    dailyCount,
-    dailyLimit,
-  }
-}
-
-/**
- * Increments both overall and daily message counters for a user.
- *
- * @param supabase - Your Supabase client.
- * @param userId - The ID of the user.
- * @param currentCounts - Current message counts (optional, will be fetchCliented if not provided)
- * @throws Error if updating fails.
- */
-export async function incrementUsage(
-  supabase: SupabaseClient,
-  userId: string,
-  currentCounts?: { messageCount: number; dailyCount: number }
-): Promise<void> {
-  let messageCount: number
-  let dailyCount: number
-
-  if (currentCounts) {
-    messageCount = currentCounts.messageCount
-    dailyCount = currentCounts.dailyCount
-  } else {
-    // If counts weren't provided, fetchClient them
-    const { data: userData, error: userDataError } = await supabase
-      .from("users")
-      .select("message_count, daily_message_count")
-      .eq("id", userId)
-      .maybeSingle()
-
-    if (userDataError || !userData) {
-      throw new Error(
-        "Error fetchClienting user data: " +
-          (userDataError?.message || "User not found")
-      )
-    }
-
-    messageCount = userData.message_count || 0
-    dailyCount = userData.daily_message_count || 0
-  }
-
-  // Increment both overall and daily message counts.
-  const newOverallCount = messageCount + 1
-  const newDailyCount = dailyCount + 1
-
-  const { error: updateError } = await supabase
-    .from("users")
-    .update({
-      message_count: newOverallCount,
-      daily_message_count: newDailyCount,
-      last_active_at: new Date().toISOString(),
-    })
-    .eq("id", userId)
-
-  if (updateError) {
-    throw new Error("Failed to update usage data: " + updateError.message)
-  }
-}
-
-/**
- * Checks the user's daily usage and increments both overall and daily counters.
- * Resets the daily counter if a new day (UTC) is detected.
- * Uses the `anonymous` flag from the user record to decide which daily limit applies.
- *
- * @param supabase - Your Supabase client.
- * @param userId - The ID of the user.
- * @throws UsageLimitError if the daily limit is reached, or a generic Error if updating fails.
- */
-export async function checkAndIncrementUsage(
-  supabase: SupabaseClient,
-  userId: string
-): Promise<void> {
-  const { userData, dailyCount } = await checkUsage(supabase, userId)
-
-  await incrementUsage(supabase, userId, {
-    messageCount: userData.message_count || 0,
-    dailyCount,
-  })
 }
 
 /**
